+++ changes.md
@@ -4,6 +4,7 @@ DartMob Changes
 **December 21, 2012
 
 * Implemented SplitLayoutPanel and Spliters. Fixed small bug in LayoutImpl. Added example.
+* Implemented FileUpload.
 
 **December 20, 2012